% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscope_down.R
\name{cytoscope_down}
\alias{cytoscope_down}
\title{download cytoscope addin by r}
\usage{
cytoscope_down(url)
}
\arguments{
\item{url}{urls by cytoscope_url()}
}
\value{
download by r
}
\description{
download cytoscope addin by r
}
\examples{
\donttest{
url <- cytoscope_url()
cytoscope_down(url)
}
}
