% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscope_url.R
\name{cytoscope_url}
\alias{cytoscope_url}
\title{Get all cytoscope addin urls}
\usage{
cytoscope_url()
}
\value{
cytoscope addin urls
}
\description{
Get all cytoscope addin urls
}
\examples{
\donttest{
url <- cytoscope_url()
}
}
