% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_series_matrix_url.R
\name{GEO_series_matrix_url}
\alias{GEO_series_matrix_url}
\title{GEO series matrix url}
\usage{
GEO_series_matrix_url(...)
}
\arguments{
\item{...}{one or more gse number}
}
\value{
url of series matrix
}
\description{
GEO series matrix url
}
\examples{
\donttest{
GEO_series_matrix_url('GSE1','GSE100')


}
}
