% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NhaNes.download.R
\name{NhaNes.download}
\alias{NhaNes.download}
\title{download NhaNes data base by r directory}
\usage{
NhaNes.download(BeginYear = NULL, yearn = 2, dirn = 1, filen = 1)
}
\arguments{
\item{BeginYear}{year of begin}

\item{yearn}{year begin number}

\item{dirn}{directory begin number}

\item{filen}{file begin number}
}
\value{
download by r directory
}
\description{
download NhaNes data base by r directory
}
\examples{
\donttest{
NhaNes.download(BeginYear = NULL, yearn=2, dirn=1, filen=1)
}
}
