% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin.R
\name{rstudio_addin}
\alias{rstudio_addin}
\title{download rstudio addin by chrome
download from https://github.com/daattali/addinslist}
\usage{
rstudio_addin(browser = getOption("browser"))
}
\arguments{
\item{browser}{browser}
}
\value{
download by web browser
}
\description{
download rstudio addin by chrome
download from https://github.com/daattali/addinslist
}
\examples{
\donttest{
rstudio_addin()
}
}
