% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_download_dir.R
\name{Target_download}
\alias{Target_download}
\title{Download Target data}
\usage{
Target_download(url, mode = "wb")
}
\arguments{
\item{url}{urls by Target_getURLs}

\item{mode}{default is wb}
}
\description{
Download Target data
}
